Extract common procedures into module.
